+++ backend/app/modules/stock/models.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     CheckConstraint,
@@ -62,9 +63,9 @@ class StockTransaction(Base, UUIDPrimaryKeyMixin):
     direction: Mapped[int] = mapped_column(SmallInteger, nullable=False)  # +1 or -1
 
     # Quantities
-    quantity_kg: Mapped[float] = mapped_column(Numeric(15, 3), nullable=False)
+    quantity_kg: Mapped[Decimal] = mapped_column(Numeric(15, 3), nullable=False)
     quantity_bags: Mapped[int | None] = mapped_column(nullable=True)
-    quantity_kip: Mapped[float | None] = mapped_column(Numeric(15, 3), nullable=True)
+    quantity_kip: Mapped[Decimal | None] = mapped_column(Numeric(15, 3), nullable=True)
     quantity_units: Mapped[int | None] = mapped_column(nullable=True)
 
     # Polymorphic source reference
